requirements.txt
androguard==3.3.5
appdirs==1.4.4
Appium-Python-Client==2.11.1
asn1crypto==1.5.1
asttokens==2.4.1
attrs==24.2.0
backports.functools-lru-cache==2.0.0
backports.shutil-get-terminal-size==1.0.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.3.0
coverage==7.6.1
cycler==0.12.1
decorator==5.1.1
dpkt==1.9.8
enum34==1.1.10
executing==2.1.0
Farama-Notifications==0.0.4
filelock==3.16.0
Flask==3.0.3
fonttools==4.53.1
frida==16.5.1
frida-tools==12.5.1
fsspec==2024.9.0
func_timeout==4.3.5
future==1.0.0
google-play-scraper==1.2.7
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
h11==0.14.0
idna==3.8
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.27.0
ipython-genutils==0.2.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
JPype1==1.5.0
kiwisolver==1.4.7
loguru==0.7.2
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.3
numpy==2.1.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathlib2==2.3.7.post1
patsy==0.5.6
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pluggy==1.5.0
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydot==3.0.1
Pygments==2.18.0
pyparsing==3.1.4
pyshark==0.6
PySocks==1.7.1
pytest==8.3.3
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
sb3_contrib==2.3.0
scandir==1.10.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
selenium==4.24.0
simplegeneric==0.8.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
stable_baselines3==2.3.2
stack-data==0.6.3
statsmodels==0.14.2
subprocess32==3.5.4
sympy==1.13.2
termcolor==2.4.0
texttable==1.7.0
threadpoolctl==3.5.0
torch==2.2.2
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
